requirements.txt
requests
colorama
dnspython
python-whois
